+++ test/services/DummyRestService_test.dart
@@ -5,6 +5,7 @@ import 'package:http/http.dart' as http;
 import 'package:pip_services3_commons/pip_services3_commons.dart';
 import '../Dummy.dart';
 import '../DummyController.dart';
+import '../SubDummy.dart';
 import './DummyRestService.dart';
 
 void main() {
@@ -52,8 +53,17 @@ void main() {
     });
 
     setUp(() {
-      _dummy1 = Dummy(id: null, key: 'Key 1', content: 'Content 1');
-      _dummy2 = Dummy(id: null, key: 'Key 2', content: 'Content 2');
+      _dummy1 = Dummy(
+          id: null,
+          key: 'Key 1',
+          content: 'Content 1',
+          array: [SubDummy(key: 'SubKey 1', content: 'SubContent 1')]);
+
+      _dummy2 = Dummy(
+          id: null,
+          key: 'Key 2',
+          content: 'Content 2',
+          array: [SubDummy(key: 'SubKey 2', content: 'SubContent 2')]);
     });
 
     test('CRUD Operations', () async {